Rename files in storage
